Mod config.org - enable word count in Doom modeline.
Comment added after testing of writeroom mode 'writeroom everywhere'
mode. The issue that can arise is that the writeroom enabled Emacs
window hides the GPG password entry box. There is no indication that
this is what has happened, it is just the case that the commit appears
to hang. Use the global writeroom with care when coding/gitting.