Mod config.org - fixed order of nodes to group configurations.
As the config file is very long (over 1000 lines at this point), I
wanted to check that the structure I had envisaged was in place. This
involved moving some nodes of the file around. Fortunately this is
very straightforward using Org Mode as you just fold everything down
to the headings, then move the headings using M-up or down to move the
nodes in relation to the nodes on the same level. You can promote or
demote nodes so that they can move relative to the other headings as
required.

The new layout makes more sense, grouping together functionality such
as 'look and feel' and 'functions that affect whole of Emacs' and
'Configurations that require installation of external packages'.

The changes did trigger an issue with the 'diminish' package which was
marked as 'require' but wasn't downloaded at that stage. Changed the
'require' to a 'use-package' and it now appears to be working again.