Mod config.org - commentary on focus mode configuration after tests.
I discovered that using my particular combination of Doom theme, Org
mode and focus mode lead to headings where the pretty bullets for org
mode headings were being shown with leading asterisks. It would appear
that this is related to the face settings not hiding these asterisks
in the expected way.

To fix this I modified to configuration for focus mode so that it
isn't activated automatically, but needs to be called using M-x
focus-mode so that it only gets used when I choose to use it.