Mod init.el file as part of configuring Povray mode and eBib.
Povray is a raytracing program, and Emacs has a mode for editing the
files required. The Povray mode includes triggers for running the
render with high and highest image quality settings, though what is
considered high quality reflects the fact that the available power was
much lower than it is now. I have modified the high/highest image
renders to both create images of the screen of my laptop (1920x1080).

The ebib configuration is just an entry for where the keywords file is
placed.